requirements.txt
discord.py==2.2.3
pyodbc==4.0.39
postgres==4.0
SQLAlchemy==2.0.15
python-dotenv==1.0.0